+++ src/Root.js
@@ -1,0 +1,43 @@
+import PropTypes from 'prop-types';
+import { Route, Switch } from 'react-router-dom';
+
+import { ROOT_ROUTE } from './constants';
+import { ConsortiumManager } from './routes';
+import ConsortiumSettings from './settings';
+
+const Root = ({
+  location,
+  match,
+  showSettings,
+}) => {
+  if (showSettings) {
+    return (
+      <ConsortiumSettings
+        location={location}
+        match={match}
+      />
+    );
+  }
+
+  return (
+    <Switch>
+      <Route
+        path={ROOT_ROUTE}
+        component={ConsortiumManager}
+      />
+    </Switch>
+  );
+};
+
+Root.propTypes = {
+  location: PropTypes.shape({
+    pathname: PropTypes.string,
+    search: PropTypes.string,
+  }),
+  match: PropTypes.shape({
+    path: PropTypes.string.isRequired,
+  }).isRequired,
+  showSettings: PropTypes.bool.isRequired,
+};
+
+export default Root;